+++ g2g-plugins/src/lib.rs
@@ -15,6 +15,7 @@ pub mod audioresample;
 pub mod audiotestsrc;
 pub mod volume;
 pub mod audiopanorama;
+pub mod audiomixer;
 pub mod capsfilter;
 pub mod fakesink;
 pub mod h264parse;